+++ app/Posts.php
@@ -4,7 +4,8 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Posts extends Model
-{
-    //
+class Posts extends Model {
+    public function upload() {
+        return $this->hasOne(Uploads::class, "post_id");
+    }
 }