+++ database/migrations/2020_11_06_192734_create_posts_table.php
@@ -14,7 +14,7 @@ class CreatePostsTable extends Migration {
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger("board_id");
-            $table->unsignedBigInteger("thread_id");
+            $table->unsignedBigInteger("thread_id")->nullable();
             $table->text("name")->nullable();
             $table->text("title")->nullable();
             $table->text("option")->nullable();