+++ database/migrations/2020_11_06_192846_create_uploads_table.php
@@ -3,6 +3,8 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\File;
+use App\Uploads;
 
 class CreateUploadsTable extends Migration {
     /**
@@ -14,9 +16,10 @@ class CreateUploadsTable extends Migration {
         Schema::create('uploads', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger("post_id");
-            $table->text("file_mime");
-            $table->text("file_location");
-            $table->text("file_size");
+            $table->text("file_name");
+            $table->text("file_path");
+            $table->text("mime_type");
+            $table->text("file_size")->nullable();
             $table->text("file_resolution")->nullable();
             $table->timestamps();
         });
@@ -28,6 +31,10 @@ class CreateUploadsTable extends Migration {
      * @return void
      */
     public function down() {
+        $uploads = Uploads::all();
+        foreach ($uploads as $u) {
+            File::delete("public/uploads/{$u->file_path}");
+        }
         Schema::dropIfExists('uploads');
     }
 }