+++ src/pages/[board]/index.svelte
@@ -14,24 +14,59 @@
         }
     }
     onDestroy(() => {
-        threads.set([]);
+        threads.set({});
     });
 </script>
 
 <style>
     .board_view {
         max-width: 1200px;
-        margin: 0 auto;
+        margin: 1rem auto;
+    }
+    .threads_list {
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(275px, 1fr));
+        grid-gap: 10px;
+    }
+    .threads_list .card img {
+        object-fit: cover;
     }
 </style>
 
 <div class="board_view">
-    <button
+    <!-- <button
         class="button is-small"
-        on:click={refresh($params.board)}>Refresh</button>
+        on:click={refresh($params.board)}>Refresh</button> -->
     <div class="threads_list">
-        {#each $threads as b}
-            <li>{b.title} - {b.content}</li>
-        {:else}Loading...{/each}
+        {#if $threads.status == 'ok'}
+            {#each $threads.data as t}
+                <div class="card">
+                    <div class="card-image">
+                        <figure class="image is-4by3">
+                            {#if t.upload}
+                                <img
+                                    src={`/uploads/${t.upload.file_path}`}
+                                    alt={t.upload.file_name} />
+                            {:else}
+                                <img
+                                    src="https://bulma.io/images/placeholders/1280x960.png"
+                                    alt="Placeholder image" />
+                            {/if}
+                        </figure>
+                    </div>
+                    <div class="card-content">
+                        <div class="media">
+                            <div class="media-content">
+                                <p class="title is-4">{t.title}</p>
+                                <p class="subtitle is-6">Anonymous</p>
+                            </div>
+                        </div>
+                        <div class="content">{t.content}</div>
+                    </div>
+                </div>
+            {:else}Loading...{/each}
+        {:else if $threads.status == 'empty'}
+            No threads found...
+        {:else}{$threads.message}{/if}
     </div>
 </div>